Collection of misc. fixes from porting to a new platform (#161)
* Fix nullptr dereference if no info_queue in TestFramework renderer
* Use std::streamoff in StateRecorder. This is the correct type for tellg/seekg
* Cast pos_type -> size_t in ReadData
* Ensure LockFreeHashMap's mBuckets is aligned to 16

+++ TestFramework/Utils/ReadData.cpp
@@ -16,7 +16,7 @@ vector<uint8> ReadData(const char *inFileName)
 	input.seekg(0, ios_base::end);
 	ifstream::pos_type length = input.tellg();
 	input.seekg(0, ios_base::beg);
-	data.resize(length);
+	data.resize(size_t(length));
 	input.read((char *)&data[0], length);
 	if (!input)
 		FatalError("Unable to read file: %s", inFileName);
